CascadeProjects/personal-website/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
PyJWT==2.10.1
requests==2.31.0
reportlab==4.0.7
openpyxl==3.1.2
qrcode==7.4.2
Pillow==10.0.0
